+++ Extension/src/background/services/settings.ts
@@ -187,6 +187,7 @@ export class SettingsService {
 
         Engine.debounceUpdate();
 
+        // TODO: Looks like not using. Maybe lost listener in refactoring.
         listeners.notifyListeners(listeners.SettingsUpdated, isImported);
         return isImported;
     }