+++ jest.config.ts
@@ -45,6 +45,8 @@ const config: Config = {
     globals: {
         // TODO: (AG-20414) Add tests for Firefox AMO
         IS_FIREFOX_AMO: false,
+        // For run tests like it's release.
+        IS_RELEASE: true,
     },
 };